requirements.txt
numpy>=1.21.0
scipy>=1.7.0
Pillow>=9.0.0
Flask>=2.0.0
Werkzeug>=2.0.0